llm-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
llama-cpp-python>=0.3.5
numpy<2.0.0
qdrant-client>=1.7.0
sentence-transformers==2.3.1
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
# chromadb>=0.5.20  # Deprecated - using Qdrant instead

# LangGraph dependencies
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5

# MinerU2.5 for advanced document parsing
--extra-index-url https://download.pytorch.org/whl/cu121
transformers>=4.30.0
torch==2.4.1+cu121
torchvision==0.19.1+cu121
pillow>=9.0.0
pdf2image>=1.16.0
accelerate>=0.20.0

